mv examples/mat_bench/run_{wrenformer=>matbench}.py
