+++ examples/mat_bench/run_matbench.py
@@ -4,65 +4,47 @@ from __future__ import annotations
 import os
 from datetime import datetime
 from os.path import dirname, isfile
-from typing import Any
+from typing import Literal
 
+import numpy as np
 import pandas as pd
 import torch
 from matbench import MatbenchBenchmark
+from matbench.task import MatbenchTask
+from torch import nn
+from torch.utils.tensorboard import SummaryWriter
 
-from aviary.roost.model import Roost
-from aviary.utils import results_multitask, train_ensemble
-from aviary.wren.model import Wren
+from aviary import ROOT
+from aviary.core import Normalizer
+from aviary.data import InMemoryDataLoader
+from aviary.losses import RobustL1Loss
+from aviary.utils import print_walltime
 from examples.mat_bench import DATA_PATHS, MatbenchDatasets
 
 __author__ = "Janosh Riebesell"
 __date__ = "2022-04-11"
 # Related Matbench issue: https://github.com/materialsproject/matbench/issues/116
 
-# %%
 torch.manual_seed(0)  # ensure reproducible results
 
-resume = False
-fine_tune = None
-transfer = None
-
-
-device = "cuda" if torch.cuda.is_available() else "cpu"
-identifiers = ("mbid", "composition")
-robust = True
-elem_emb = "matscholar200"
-
-ensemble = 1
-run_id = 1
-log = True
-
-data_params: dict[str, Any] = {
-    "batch_size": 128,
-    "num_workers": 0,
-    "pin_memory": False,
-    "shuffle": True,
-}
-setup_params = {
-    "optim": "AdamW",
-    "learning_rate": 3e-4,
-    "weight_decay": 1e-6,
-    "momentum": 0.9,
-    "device": device,
-}
-restart_params = {
-    "resume": resume,
-    "fine_tune": fine_tune,
-    "transfer": transfer,
-}
-
 
 # %%
+learning_rate = 1e-3
+warmup_steps = 10
+
+
+def lr_lambda(epoch: int) -> float:
+    return min((epoch + 1) ** (-0.5), (epoch + 1) * warmup_steps ** (-1.5))
+
+
+@print_walltime("run_matbench_task()")
 def run_matbench_task(
     model_name: str,
     benchmark_path: str,
     dataset_name: MatbenchDatasets,
+    fold: Literal[0, 1, 2, 3, 4],
     epochs: int = 100,
-) -> MatbenchBenchmark | None:
+) -> MatbenchBenchmark:
     """Run a single matbench task.
 
     Args:
@@ -79,12 +61,22 @@ def run_matbench_task(
     Returns:
         MatbenchBenchmark: Benchmark results.
     """
-    if "roost" in model_name.lower():
-        model_cls = Roost
+    if "wrenformer" in model_name.lower():
+        from aviary.wrenformer.data import (
+            collate_batch,
+            wyckoff_embedding_from_aflow_str,
+        )
+        from aviary.wrenformer.model import Wrenformer as ModelClass
+    # TODO: make it work with Wren and Roost too at some point, currently the model and
+    # data class kwargs are differently named
     elif "wren" in model_name.lower():
-        model_cls = Wren
+        from aviary.wren.data import collate_batch
+        from aviary.wren.model import Wren as ModelClass
+    elif "roost" in model_name.lower():
+        from aviary.roost.data import collate_batch
+        from aviary.roost.model import Roost as ModelClass
     else:
-        raise ValueError(f"{model_name = } must contain either 'wren' or 'roost'")
+        raise ValueError(f"Unexpected {model_name = }")
 
     if not benchmark_path.endswith(".json"):
         raise ValueError(f"{benchmark_path = } must have .json extension")
@@ -93,133 +85,114 @@ def run_matbench_task(
     else:
         mbbm = MatbenchBenchmark(subset=[dataset_name])
 
-    task = mbbm.tasks_map[dataset_name]
+    matbench_task: MatbenchTask = mbbm.tasks_map[dataset_name]
 
-    if task.all_folds_recorded:
-        print(f"\nTask {dataset_name} already recorded! Skipping...\n")
-        return None
+    if matbench_task.is_recorded[fold]:
+        print(f"{fold = } of {dataset_name} already recorded! Skipping...")
+        return mbbm
 
-    df = pd.read_json(DATA_PATHS[dataset_name])
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    print(f"Pytorch running on {device=}")
 
-    target, task_type = (str(task.metadata[x]) for x in ("target", "task_type"))
-    task_dict = {target: task_type}
-    loss_dict = {target: "L1" if task_type == "regression" else "CSE"}
+    df = pd.read_json(DATA_PATHS[dataset_name]).set_index("mbid", drop=False)
+    with print_walltime("Generating initial Wyckoff embedding"):
+        df["features"] = df.wyckoff.map(wyckoff_embedding_from_aflow_str)
+    matbench_task.df = df
 
-    if model_cls == Roost:
-        from aviary.roost.data import CompositionData as data_cls
-        from aviary.roost.data import collate_batch
-    else:
-        from aviary.wren.data import WyckoffData as data_cls
-        from aviary.wren.data import collate_batch
-        from aviary.wren.utils import count_wyks
+    target, task_type = (
+        str(matbench_task.metadata[x]) for x in ("target", "task_type")
+    )
+    task_dict = {target: task_type}  # e.g. {'exfoliation_en': 'regression'}
 
-    task.df = df.set_index("mbid", drop=False)
-    data_params["collate_fn"] = collate_batch
+    robust = False
+    loss_func = (
+        (RobustL1Loss if robust else nn.L1Loss())
+        if task_type == "regression"
+        else (nn.NLLLoss() if robust else nn.CrossEntropyLoss())
+    )
+    loss_dict = {target: (task_type, loss_func)}
+    normalizer_dict = {target: Normalizer() if task_type == "regression" else None}
 
-    for fold in task.folds:
-        if task.is_recorded[fold]:
-            print(f"{fold = } of {dataset_name} already recorded! Skipping...")
-            continue
+    fold_name = f"{model_name}-{dataset_name}-{fold=}"
 
-        model_name = f"roost-{dataset_name}-{fold=}"
+    print(f"{matbench_task=}")
+    train_df = matbench_task.get_train_and_val_data(fold, as_type="df")
+    test_df = matbench_task.get_test_data(fold, as_type="df", include_target=True)
 
-        train_df = task.get_train_and_val_data(fold, as_type="df")
-        test_df = task.get_test_data(fold, as_type="df", include_target=True)
+    features, targets, ids = (train_df[x] for x in ["features", target, "mbid"])
+    targets = torch.tensor(targets, device=device)
+    features_arr = np.empty(len(features), dtype=object)
+    for idx, tensor in enumerate(features):
+        features_arr[idx] = tensor.to(device)
 
-        # need to restrict max number of Wyckoff positions to avoid OOM errors
-        # (even on 80 GB A100 GPUs)
-        if model_cls == Wren:
-            train_df = train_df[[count_wyks(x) < 20 for x in train_df.wyckoff]]
+    train_loader = InMemoryDataLoader(
+        [features_arr, targets, ids],
+        batch_size=32,
+        shuffle=True,
+        collate_fn=collate_batch,
+    )
 
-        train_set = data_cls(train_df, task_dict, elem_emb, identifiers=identifiers)
+    features, targets, ids = (test_df[x] for x in ["features", target, "mbid"])
+    targets = torch.tensor(targets, device=device)
+    features_arr = np.empty(len(features), dtype=object)
+    for idx, tensor in enumerate(features):
+        features_arr[idx] = tensor.to(device)
 
-        test_set = data_cls(test_df, task_dict, elem_emb, identifiers=identifiers)
+    test_loader = InMemoryDataLoader(
+        [features_arr, targets, ids], batch_size=1024, collate_fn=collate_batch
+    )
 
-        model_params = {
-            "task_dict": task_dict,
-            "robust": robust,
-            "n_targets": train_set.n_targets,
-            "elem_emb_len": train_set.elem_emb_len,
-            "elem_fea_len": 64,
-            "n_graph": 3,
-            "elem_heads": 3,
-            "elem_gate": [256],
-            "elem_msg": [256],
-            "cry_heads": 3,
-            "cry_gate": [256],
-            "cry_msg": [256],
-            "trunk_hidden": [128, 128],
-            "out_hidden": [64, 64],
-        }
+    # n_features = element + wyckoff embedding lengths + element weights in composition
+    model = ModelClass(
+        n_targets=[1], n_features=200 + 444 + 1, task_dict=task_dict, robust=robust
+    )
+    model.to(device)
+    optimizer = torch.optim.AdamW(params=model.parameters(), lr=learning_rate)
 
-        if model_cls == Wren:
-            model_params["sym_emb_len"] = train_set.sym_emb_len
+    scheduler = torch.optim.lr_scheduler.LambdaLR(optimizer, lr_lambda, verbose=True)
 
-        train_ensemble(
-            model_class=model_cls,
-            model_name=model_name,
-            run_id=run_id,
-            ensemble_folds=ensemble,
-            epochs=epochs,
-            train_set=train_set,
-            val_set=test_set,
-            log=log,
-            data_params=data_params,
-            setup_params=setup_params,
-            restart_params=restart_params,
-            model_params=model_params,
-            loss_dict=loss_dict,
-        )
+    writer = SummaryWriter(f"{ROOT}/runs/{fold_name}/{datetime.now():%Y-%m-%d@%H-%M}")
 
-        results_dict = results_multitask(
-            model_class=model_cls,
-            model_name=model_name,
-            run_id=run_id,
-            ensemble_folds=ensemble,
-            test_set=test_set,
-            # we set shuffle=False since ensembling needs fixed data order
-            data_params={**data_params, "shuffle": False},
-            robust=robust,
-            task_dict=task_dict,
-            device=device,
-            eval_type="checkpoint",
-            save_results=False,
-        )
+    model.fit(
+        train_loader,
+        test_loader,
+        optimizer,
+        scheduler,
+        epochs=epochs,
+        criterion_dict=loss_dict,
+        normalizer_dict=normalizer_dict,
+        model_name=fold_name,
+        run_id=1,
+        checkpoint=False,
+        writer=writer,
+    )
 
-        if task_type == "regression":
-            predictions = results_dict[target]["pred"].ravel()
-        else:  # classification
-            logits = results_dict[target]["logits"].ravel()
-            predictions = logits.argmax(axis=1)
+    _, [predictions], *_ = model.predict(test_loader)
 
-        # record model predictions
-        task.record(fold, predictions)
+    if isfile(benchmark_path):  # we checked for isfile() above but possible another
+        # slurm job created a partial benchmark in the meantime in which case we merge results
+        mbbm = MatbenchBenchmark.from_file(benchmark_path)
+        matbench_task = mbbm.tasks_map[dataset_name]
+        # task needs df to pass validation before fold can be recorded
+        matbench_task.df = df
+    elif benchmark_dir := dirname(benchmark_path):
+        os.makedirs(benchmark_dir, exist_ok=True)
+
+    # record model predictions
+    matbench_task.record(fold, predictions.cpu())
+    mbbm.tasks_map[dataset_name] = matbench_task
 
     # save model benchmark
-    if isfile(benchmark_path):  # we checked for isfile() above but possible another
-        # slurm job created it in the meantime in which case we need to merge results
-        mbbm = MatbenchBenchmark.from_file(benchmark_path)
-        mbbm.tasks_map[dataset_name] = task
-    else:
-        os.makedirs(dirname(benchmark_path), exist_ok=True)
-
     mbbm.to_file(benchmark_path)
 
     return mbbm
 
 
+# %%
 if __name__ == "__main__":
-    from argparse import ArgumentParser
-
-    parser = ArgumentParser()
-    parser.add_argument("model", choices=["roost", "wren"])
-    args = parser.parse_args()
-
-    mbbm = MatbenchBenchmark()
-    benchmark_path = f"{args.model}-matbench-{datetime.now():%Y-%m-%d@%H-%M}.json"
-
-    model_cls = Roost if args.model == "roost" else Wren
-
-    for idx, dataset_name in enumerate(DATA_PATHS, 1):
-        print(f"\n\n{idx}/{len(DATA_PATHS)}")
-        run_matbench_task(dataset_name, benchmark_path, args.model, model_cls)
+    # for testing and debugging
+    model_name = "wrenformer"
+    benchmark_path = "wrenformer-tmp.json"
+    dataset = "matbench_jdft2d"
+    # dataset = "matbench_mp_is_metal"
+    run_matbench_task(model_name, benchmark_path, dataset, 4, epochs=1)
